requirements.txt
fastapi
uvicorn 
anthropic
python-dotenv
